Move inner SO class to new IdsStreamingOutput
